seperating get_user into its own module
